refactor(Routes & Action creators):  Work under construction 🚧
- Company profile refactorized to use token and new actions files.
