+++ client/src/components/Dashboard/DashboardNavbar.tsx
@@ -31,7 +31,10 @@ const DashboardNavbar: FC = () => {
   return (
     <Container>
       <ButtonsContainer>
-
+        <Notifications role={role} />
+        {role === "applicant" && (
+          <Favorites role={role} />
+        )}
         <ProfModalButton user={user} />
       </ButtonsContainer>
     </Container>
@@ -39,8 +42,3 @@ const DashboardNavbar: FC = () => {
 };
 
 export default DashboardNavbar;
-
-        //<Notifications role={role} />
-        //{role === "applicant" && (
-          //<Favorites role={role} />
-        //)}
