+++ client/src/components/EditCompanyProfile/EditCompanyProfile.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import {useEffect, useState} from "react";
 import {
     CardContent,
     Input,
@@ -12,24 +12,26 @@ import {
     CardHeader,
     EditButton,
 } from "./Styles";
-import { useDispatch, useSelector } from "react-redux";
-import { useParams } from "react-router";
-import { editCompany, getCompany } from "../../redux/actions/actionCreators";
+import {useDispatch, useSelector} from "react-redux";
+import {useParams} from "react-router";
+import {getCompany} from "../../redux/actions/public/generalActions";
+import {editCompany} from "../../redux/actions/private/companyActions";
 import Dashboard from "../../pages/Dashboard/Dashboard";
-import { AboutCompany } from "./AboutCompany";
-import { Mission } from "./Mission";
-import { Vission } from "./Vission";
-import { Values } from "./Values";
+import {AboutCompany} from "./AboutCompany";
+import {Mission} from "./Mission";
+import {Vission} from "./Vission";
+import {Values} from "./Values";
 
 export const CompanyProfile = () => {
     const dispatch = useDispatch();
-    const { companyId } = useParams();
-    const company = useSelector(
-        (state: any) => state.companyReducer.companyDetail
-    );
+    const {companyId} = useParams();
     useEffect(() => {
         dispatch(getCompany(companyId));
     }, [dispatch, companyId]);
+    const company = useSelector(
+        (state: any) => state.companyReducer.companyDetail
+    );
+    console.log("Company: ", company)
 
     const [companyInfo, setCompanyInfo] = useState(company);
     const [isEdit, setIsEdit] = useState(false);
