+++ client/src/components/Menus/PrivateMenu.tsx
@@ -111,6 +111,10 @@ const PrivateMenu: FC<Props> = ({role}) => {
         )}
       </NavLinks>
       <ButtonsContainer>
+        <Notifications role={role} />
+        {role === "applicant" && (
+          <Favorites role={role} />
+        )}
         <ProfModalButton user={user} />
       </ButtonsContainer>
     </>
@@ -118,8 +122,3 @@ const PrivateMenu: FC<Props> = ({role}) => {
 };
 
 export default PrivateMenu;
-
-        //<Notifications role={role} />
-        //{role === "applicant" && (
-          //<Favorites role={role} />
-        //)}
