+++ client/src/redux/actions/private/companyActions.tsx
@@ -4,10 +4,10 @@ import Storage from "../../../services/storage";
 import {ActionType} from "../actionTypes";
 import {Action} from "../index";
 
-
+const token = Storage.get("token");
 
 export const createPost =
-  (dataPost: any, token: string) => async (dispatch: Dispatch<Action>) => {
+  (dataPost: any, tokenId: string) => async (dispatch: Dispatch<Action>) => {
     try {
       console.log("Data enviada: ", dataPost);
       console.log(token);
@@ -17,7 +17,7 @@ export const createPost =
         dataPost,
         {
           headers: {
-            token: token,
+            token: tokenId,
           },
         }
       );
@@ -56,7 +56,13 @@ export const editPost =
     try {
       let {data} = await axios.put(`/posts/update/${postId}`, {
         endDate: endDate,
-      });
+      },
+        {
+          headers: {
+            token: token || "",
+          },
+        }
+      );
       //console.log(data);
       console.log("POSTS", "Se recibe data de la API");
       return dispatch({type: ActionType.EDIT_POST, payload: data});
@@ -71,7 +77,13 @@ export const setFavApplicant =
       let {data} = await axios.put(`/company/favorites`, {
         applicantId,
         postId
-      });
+      },
+        {
+          headers: {
+            token: token || "",
+          },
+        }
+      );
       return dispatch({type: ActionType.SET_FAV_APPLICANT, payload: data.favorites});
     } catch (error) {
       console.log(error);
@@ -82,7 +94,13 @@ export const setFavApplicant =
 export const sendMercadoPago = (payload: any) => {
   return async function () {
     try {
-      var response = await axios.post("http://localhost:3002/api/v2/payment/checkout", payload)
+      var response = await axios.post("http://localhost:3002/api/v2/payment/checkout", payload,
+        {
+          headers: {
+            token: token || "",
+          },
+        }
+      )
       let resp = response.data.init_point;
       window.open(resp);
     } catch (error) {
@@ -94,7 +112,13 @@ export const sendMercadoPago = (payload: any) => {
 export const updateInfo = (payload: any) => {
   return async function () {
     try {
-      var response = await axios.post("http://localhost:3002/api/v2/payment/checkout", payload)
+      var response = await axios.post("http://localhost:3002/api/v2/payment/checkout", payload,
+        {
+          headers: {
+            token: token || "",
+          },
+        }
+      )
       console.log('action updateInfo: ', response)
       return response;
 
@@ -103,3 +127,29 @@ export const updateInfo = (payload: any) => {
     }
   }
 }
+
+export const editCompany =
+  (companyData: object, companyId: number) =>
+    async (dispatch: Dispatch<Action>) => {
+      console.log("Id: ", companyId)
+      console.log("Data enviada: ", companyData);
+      try {
+        const {data} = await axios.put(
+          `/company/update/${companyId}`,
+          companyData,
+          {
+            headers: {
+              token: token || "",
+            },
+          }
+        );
+        //console.log("Data recibida: ", data);
+        console.log("Información actualizada");
+        return dispatch({
+          type: ActionType.EDIT_COMPANY,
+          payload: data,
+        });
+      } catch (error) {
+        console.log(error);
+      }
+    };
